Refactor: drive 모듈 구조 변경 - drive 내 folder, file, document 요소 통합
